third_party/WebKit/LayoutTests/platform/android/fast/forms/time-multiple-fields/time-multiple-fields-change-layout-by-value-expected.txt

Change multiple fields time input UI layout by value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL widthOfStep60 < widthWithMillisecond should be true. Was false.
PASS widthOfStep60 == widthWithSecond is true
PASS successfullyParsed is true

TEST COMPLETE
